pyproject.toml
[tool.poetry]
name = "communityvibe"
version = "0.1.0"
description = "CommunityVibe - Social Media Analytics Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
pydantic = "^2.6.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
praw = "^7.7.1"
tweepy = "^4.14.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
transformers = "^4.37.2"
apscheduler = "^3.10.4"
pydantic-settings = "^2.1.0"
openai = "^1.12.0"
tenacity = "^8.2.3"
email-validator = "^2.1.0.post1"
celery = "^5.3.6"
redis = "^5.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 